refactor: split plugin interactive registration
